requirements.txt
aiogram==2.25.1
python-dotenv==0.21.1
SQLAlchemy==1.4.41
Pillow==9.5.0
requests==2.28.2
python-telegram-bot-pagination==1.0.0 